Permissions page creation

+++ sick-fits/backend/src/resolvers/Mutation.js
@@ -3,10 +3,11 @@ const { randomBytes } = require('crypto');
 const jwt = require('jsonwebtoken');
 const { promisify } = require('util');
 const { transport, makeANiceEmail } = require('../mail');
+const { hasPermission } = require('../utils');
 
 const Mutation = {
     async createItem(parent, args, ctx, info) {
-        if(ctx.request.userId) {
+        if(!ctx.request.userId) {
             throw new Error('You must be logged in.');
         }
 
@@ -38,10 +39,21 @@ const Mutation = {
     },
 
     async deleteItem(parent, args, ctx, info) {
+        // throw new Error('You aren\'t allowed to delete this item');
+
         const where = {id: args.id};
         // Find the item
-        const item = await ctx.db.query.item({where}, `{id  title}`);
+        const item =
+         await ctx.db.query.item({where}, `{id  title user { id }}`);
         // TODO: Check if the user owns that item or have permissions
+        const ownsItem = item.user.id === ctx.request.userId;
+        const hasPermissions = 
+                ctx.request.user.permissions.some(
+                    permission => ['ADMIN', 'ITEMDELETE'].includes(permission));
+        if (!ownsItem && !hasPermissions) {
+            throw new Error('You aren\'t allowed to delete this item');
+        }
+
 
         // Delete item from DB
         return ctx.db.mutation.deleteItem({where}, info);
@@ -153,9 +165,39 @@ const Mutation = {
         });
 
         return modifiedUser;
+    },
+
+    async updatePermissions(parent, args, ctx, info) {
+        // 1. Check if they are logged in
+        if (!ctx.request.userId) {
+            throw new Error('You must be logged in!');
+        }
+        // 2. Query the current user
+        const currentUser = await ctx.db.query.user(
+            {
+                where: {
+                    id: ctx.request.userId,
+                },
+            },
+            info
+        );
+        // 3. Check if they have permissions to do this
+        hasPermission(currentUser, ['ADMIN', 'PERMISSIONUPDATE']);
+        // 4. Update the permissions
+        return ctx.db.mutation.updateUser(
+            {
+                data: {
+                    permissions: {
+                        set: args.permissions,
+                    },
+                },
+                where: {
+                    id: args.userId,
+                },
+            },
+            info
+        );
     }
-
-
 };
 
 module.exports = Mutation;
